홈페이지 로직 구현
